OnlineBoutiqueAgent/ecommerce_agent/requirements.txt
google-adk==1.0.0
google-generativeai==0.8.5
google-genai

mcp==1.9.1
requests==2.32.3
beautifulsoup4==4.13.4

pillow==12.1.0
python-dotenv==1.1.0

flask==3.0.0
httpx==0.28.1
deprecated


pydantic>=2.0.0
fastapi>=0.109.0
uvicorn>=0.27.0

reportlab>=4.0.0
pymongo>=4.6.0
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0

numpy>=1.24.0

# IMPORTANT — pinned versions
mediapipe==0.10.9
opencv-python==4.13.0.92
opencv-contrib-python==4.13.0.92

rembg>=2.0.0
